requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.1